This commit tries to fix two issues: #47 and #38 else statement that was supposed to highlight settext style headers.
The partial solution for #38 solves the block highlighting problems that @mjnaderl pointed out.
What was causing this was not checking if a block was already completed
on the same line it was written at and not continuing the previous block format.

Similar fixes still need to be done for the inline patterns (**, __, //, etc.), which will come later.