Convenience .local.vimrc example for notebooks
This .local.vimrc allows easily navigating your notebook with
the gf in vim. You can source this manually (after placing it
in the root of your notes folder of the notebook) or just use
the vim-localrc plugin to source it automatically whenever
viewing a mikidown file in vim. Further tweaks that could be
made include, for example, supporting TOC reference opening
through gf without changing any of the mikidown syntax.