Don't assume that gf was done on a wiki page
This is for the convenience .local.vimrc (which is
NOT the project settings to make sure all indents are
spaces), which makes browsing a mikidown notebook from
vim a lot easier in cases where one's trapped on
the commandline.

In the cases where it isn't and isn't an https? link,
execute the normal gf behavior. There probably needs to
be some sort of consideration for those https links though.