Start working on templating system
This is for #54. As a nice bonus, that means the default template
can be translated instead of being hard coded in English. Also,
the command title template type is commented out because I'm not
too sure how useful that'd be to have as a template title type
(on top of possible security concerns like... someone putting
rm in the title template name. So it's safer just to have
FSTRING and DATETIME for now).