Decouple MikiWindow from MikiTray
Addresses #75

To do this in the least intrusive manner for every MikiWindow instance,
two class-level events were added to MikiWindow.

Then, MikiTray is registered as a listener for both post creation/open
and post close. This allows MikiTray to keep track of all of the
MikiWindows open without changing the function signatures that much.

There is one quirk about the implementation, though.

If the window is on another desktop, it won't try to place to place
the window in a place where it can be seen. This is true even before
the decoupling. Workaround for this is double clicking the menu
entry so the window is on the correct desktop.