Implement minimize to system tray
Addresses #75

To do this, the MikiTray sets the tray icon attribute to itself when it
listens to a new window being initialized.

Other than the window hiding, there's another attribute that needed to
be added to MikiWindow: alwaysClose.

alwaysClose defaults to false except if the user explicitly quits it.
This is so the code that detects minimize to system tray possibilities
does not interfere with normal closing of a notebook.