__version__ as tuple for better version checks
If left alone at a string, the comparison is left at the mercy of the
normal string sort order (which is alphabetically based). Explicitly
stating the components of the version allows for more explicit version
comparisons.

It also makes it more clear when a version bump is major, minor, or
just a patch.